Finished with step 7. merging all cut videos.

+++ src/index.ts
@@ -4,23 +4,28 @@ import { ReadableStream } from "web-streams-polyfill";
 global.ReadableStream = ReadableStream;
 
 import {
+  createFolder,
   isFileExist,
   readAllVideos,
   readTextFile,
   writeTextFile,
 } from "./utils/fs-utils";
-import { convertToSound } from "./utils/ffmpeg-utils";
+import { convertToSound, cutVideo, mergeVideos } from "./utils/ffmpeg-utils";
 import { storeTranscript } from "./utils/ai-utils";
 import { summaryChain } from "./chains/summary_chain";
+import { gettimeCutsFromTranscript } from "./utils/transcript.util";
 
 const videosRoot = path.join(__dirname, "../videos");
 
 async function main() {
+  //
   // 1. Reading a list of videos
+  //
   console.log("Reading a list of videos");
   const allVideos = readAllVideos();
 
   for (let i = 0; i < allVideos.length; i++) {
+    const videoName = path.basename(allVideos[i].split(".")[0]);
     const videoPath = path.join(videosRoot, allVideos[i]);
 
     const soundOutputFileName = path.join(
@@ -28,8 +33,10 @@ async function main() {
       allVideos[i].split(".")[0] + ".mp3"
     );
 
+    //
     // 2. Convert each video to audio
-    if (isFileExist(soundOutputFileName)) {
+    //
+    if (!isFileExist(soundOutputFileName)) {
       console.log("Converting video to audio", soundOutputFileName);
       await convertToSound(videoPath, soundOutputFileName);
     }
@@ -39,26 +46,67 @@ async function main() {
       allVideos[i].split(".")[0] + ".srt"
     );
 
+    //
     // 3. Get transcript for each audio
+    //
     if (!isFileExist(transcriptFileName)) {
       console.log("Getting transcript for audio", soundOutputFileName);
       await storeTranscript(soundOutputFileName, transcriptFileName, "fa");
     }
 
-    // 5. Ask AI to generate a summary all transcripts and return it as a text file, and keeping the time data.
+    //
+    // 4. Ask AI to generate a summary all transcripts and return it as a text file, and keeping the time data.
+    //
     const transcript = readTextFile(transcriptFileName);
-    console.log("Getting transcript summary for", transcriptFileName);
-    await summaryChain.invoke({ transcript }).then((summary) => {
-      const summaryFileName = path.join(
-        videosRoot,
-        allVideos[i].split(".")[0] + ".summary.srt"
+
+    const summaryFileName = path.join(
+      videosRoot,
+      allVideos[i].split(".")[0] + ".summary.srt"
+    );
+
+    if (!isFileExist(summaryFileName)) {
+      console.log("Getting transcript summary for", transcriptFileName);
+      await summaryChain.invoke({ transcript }).then((summary) => {
+        writeTextFile(summaryFileName, summary.text);
+      });
+    }
+
+    // 5. cute out the video based on the summary time data.
+    // -- ffmpeg
+    const timeCuts = gettimeCutsFromTranscript(readTextFile(summaryFileName));
+
+    // 6. Cutout videos based on the summary time data
+    for (let timeIndex = 0; timeIndex < timeCuts.length; timeIndex++) {
+      const timeCute = timeCuts[timeIndex];
+
+      const cutDirectory = path.join(videosRoot, videoName);
+      createFolder(cutDirectory);
+
+      const cutFileName = path.join(
+        cutDirectory,
+        `${videoName}-${timeIndex}.mp4`
       );
 
-      writeTextFile(summaryFileName, summary.text);
-    });
+      if (!isFileExist(cutFileName)) {
+        console.log("Cut video", videoName, timeCute);
+        await cutVideo(videoPath, cutFileName, timeCute.start, timeCute.end);
+      }
+    }
+
+    // 7. Merge all cut videos
+    const cutVideos = readAllVideos(path.join(videosRoot, videoName)).map(
+      (video) => path.join(videosRoot, videoName, video)
+    );
+
+    const mergePath = path.join(videosRoot, videoName + "_short.mp4");
+
+    if (!isFileExist(mergePath)) {
+      console.log("Merging videos", cutVideos);
+      await mergeVideos(cutVideos, mergePath);
+    }
   }
 
-  // 6. Cutout videos based on the summary time data and combine them.
+  //  and combine them.
   // -- ffmpeg
 }
